Fit all 5+2 layers into a 300x600mm sheet
Fit all the layers needed for an optimal assembly with the Adafruit
USB-C breakout board. The spacing layer is made of two patterns that
account for 1) the whole PCB size, and 2) the USB-C socket mounted
on it.  Every layer (besides the one enclosing the PCB) feature
two additional M2 screws holes to fasten the PCB.

There are two extra "full" spacers (without a hole to let a cable
through) broken into four parts each. These are there as buffer if
an extra layer is needed because of a thick wiring layer or if
something bad were to happen to another spacer.